// packages/uimodel-composer/src/generated/css/StyleRule.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, EClass, BasicEObject, EStructuralFeature } from '@emfts/core';
import type { Expression } from '..';
import type { CssDeclaration } from './CssDeclaration';
import type { CssState } from './CssState';

/**
 * StyleRule
 * @generated
 */
export interface StyleRule extends EObject {
  name?: string;
  targetClass?: EClass;
  includeSubtypes?: boolean;
  componentName?: string;
  group?: string;
  priority?: number;
  media?: string;
  condition?: Expression;
  declarations: CssDeclaration[];
  states: CssState[];

}